refactor(Result): delete function call
For smooth unit test progress and separation of functions,
it is necessary to call a function that is executed at the same time as the instance is created to be executed.

BREAKING CHANGE: Delete function call when instance created.

Before:

constructor(...){
// ... //
this.calculateEachLotto();
}

After:

constructor(...){
// ... //
}

Delete function call in the constructor.