refactor(Lotto): modify the code structure
As the parameters of the constructor are modified, the functions declared in the class are modified and deleted.
This commit is a modification to satisfy the mission restriction that only one parameter must be entered in the constructor of the ‘Lotto’ class.

BREAKING CHANGE: Modify the code structure.

Before:

3 parameters required in constructor of class ‘Lotto’
Console.readLine to input bonus number exists in ‘Lotto’ class.

After:

Only one parameter is required in the constructor.

Bonus number input function removed.