refactor(App): change exception handling
In the past, exception handling was carried out only for points, but other exceptions could not be dealt with.
So use regular expressions and change to stricter exception handling so that only numeric characters can be entered.

BREAKING CHANGE: Change exception handling

Before:

Exception handling only for points

After:

Exception handling occurs for all input values that are not numeric characters